requirements.txt
# Trading Bot Requirements

yfinance
pandas
numpy
requests
python-dotenv
alpaca-trade-api
alpaca-py
websockets